Closes trac:903: MX BTL fix
Submitted by bosilca, Reviewed by Tim Mattox & Galen, RM-Approved on telecon

r12517: Indentation only.

r13423: No more warnings from the datatype engine.

r13630:
Correctly handle the pack and unpack for contiguous types with gaps.
The pack/unpack leaves the convertor in a consistent state, such
that the next operation will succeed.

This commit was SVN r13646.

The following SVN revision numbers were found above:
  r12517 --> open-mpi/ompi@da1720d1ef70388879f56758ee524c5af7e405af
  r13423 --> open-mpi/ompi@f3a5fe488ebf7d478fd57194da71065bcfe7eae9
  r13630 --> open-mpi/ompi@5214e3751bb5b47304eb60af2c2f08492a834651

The following Trac tickets were found above:
  Ticket 903 --> https://svn.open-mpi.org/trac/ompi/ticket/903

+++ ompi/datatype/datatype_unpack.h
@@ -36,8 +36,8 @@ static inline void unpack_predefined_data( ompi_convertor_t* CONVERTOR, /* the c
         /* the extent and the size of the basic datatype are equals */
         OMPI_DDT_SAFEGUARD_POINTER( _destination, _copy_blength, (CONVERTOR)->pBaseBuf,
                                     (CONVERTOR)->pDesc, (CONVERTOR)->count );
-        DO_DEBUG( opal_output( 0, "unpack 1. memcpy( %p, %p, %ld ) => space %d\n",
-                               _destination, *(SOURCE), _copy_blength, *(SPACE) ); );
+        DO_DEBUG( opal_output( 0, "unpack 1. memcpy( %p, %p, %ld ) => space %lu\n",
+                               _destination, *(SOURCE), _copy_blength, (unsigned long)(*(SPACE)) ); );
         MEMCPY_CSUM( _destination, *(SOURCE), _copy_blength, (CONVERTOR) );
         *(SOURCE)    += _copy_blength;
         _destination += _copy_blength;
@@ -46,8 +46,8 @@ static inline void unpack_predefined_data( ompi_convertor_t* CONVERTOR, /* the c
         for( _i = 0; _i < _copy_count; _i++ ) {
             OMPI_DDT_SAFEGUARD_POINTER( _destination, _copy_blength, (CONVERTOR)->pBaseBuf,
                                         (CONVERTOR)->pDesc, (CONVERTOR)->count );
-            DO_DEBUG( opal_output( 0, "unpack 2. memcpy( %p, %p, %ld ) => space %d\n",
-                                   _destination, *(SOURCE), _copy_blength, *(SPACE) - (_i * _copy_blength) ); );
+            DO_DEBUG( opal_output( 0, "unpack 2. memcpy( %p, %p, %ld ) => space %lu\n",
+                                   _destination, *(SOURCE), _copy_blength, (unsigned long)(*(SPACE) - (_i * _copy_blength)) ); );
             MEMCPY_CSUM( _destination, *(SOURCE), _copy_blength, (CONVERTOR) );
             *(SOURCE)    += _copy_blength;
             _destination += _elem->extent;
